done ui task and starting firebase auth task
