--- lib/pages/bottomnav.dart
@@ -1,51 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:intern_task/pages/home_screen.dart';
-import 'package:intern_task/pages/myaccount_screen.dart';
-import 'package:intern_task/pages/myclasses_screen.dart';
-
-class BottomNav extends StatefulWidget {
-  const BottomNav({super.key});
-
-  @override
-  State<BottomNav> createState() => _BottomNavState();
-}
-
-class _BottomNavState extends State<BottomNav> {
-  int _selectedIndex = 0;
-  final screens = [
-    const HomeScreen(),
-    const MyClassesScreen(),
-    const MyAccountScreen(),
-  ];
-
-  @override
-  Widget build(BuildContext context) {
-    return SafeArea(
-      child: Scaffold(
-        body: IndexedStack(
-          index: _selectedIndex,
-          children: screens,
-        ),
-        bottomNavigationBar: BottomNavigationBar(
-          items: const [
-            BottomNavigationBarItem(
-              icon: Icon(Icons.home),
-              label: 'Home',
-            ),
-            BottomNavigationBarItem(
-              icon: Icon(Icons.book),
-              label: 'My Classes',
-            ),
-            BottomNavigationBarItem(
-              icon: Icon(Icons.person),
-              label: 'Account',
-            ),
-          ],
-          currentIndex: _selectedIndex,
-          selectedItemColor: Theme.of(context).primaryColor,
-          onTap: (index) => setState(() => _selectedIndex = index),
-        ),
-      ),
-    );
-  }
-}
