--- lib/pages/menu_button.dart
@@ -1,28 +1,0 @@
-import 'package:flutter/material.dart';
-
-class MenuButton extends StatelessWidget {
-  final String buttonName;
-
-  const MenuButton({super.key, required this.buttonName});
-
-  @override
-  Widget build(BuildContext context) {
-    return InkWell(
-      onTap: () {},
-      highlightColor: Theme.of(context).primaryColor,
-      child: Container(
-        padding: const EdgeInsets.all(10),
-        child: Row(
-          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-          children: [
-            Text(
-              buttonName,
-              style: Theme.of(context).textTheme.headline6,
-            ),
-            const Icon(Icons.arrow_forward_ios)
-          ],
-        ),
-      ),
-    );
-  }
-}
