+++ lib/screens/bookclass_screen.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+class BookClassScreen extends StatefulWidget {
+  const BookClassScreen({super.key});
+
+  @override
+  State<BookClassScreen> createState() => _BookClassScreenState();
+}
+
+class _BookClassScreenState extends State<BookClassScreen> {
+  @override
+  Widget build(BuildContext context) {
+    return Container();
+  }
+}
