+++ lib/screens/home_screen.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:intern_task/pages/course_card.dart';
+import 'package:intern_task/util/widget/card/course_card_list.dart';
 
 class HomeScreen extends StatefulWidget {
   const HomeScreen({super.key});
@@ -16,11 +16,14 @@ class _HomeScreenState extends State<HomeScreen> {
         appBar: AppBar(
             title: const Text('Home'),
             leading: IconButton(
-                onPressed: () {Navigator.pop(context);},
+                onPressed: () {
+                  Navigator.pop(context);
+                },
                 icon: const Icon(Icons.arrow_back_ios_sharp))),
-        body: Container(
-          padding: const EdgeInsets.symmetric(horizontal: 30),
+        body: Padding(
+          padding: const EdgeInsets.symmetric(horizontal: 15.0),
           child: Column(
+            mainAxisSize: MainAxisSize.min,
             crossAxisAlignment: CrossAxisAlignment.start,
             children: [
               const SizedBox(height: 30),
@@ -31,6 +34,7 @@ class _HomeScreenState extends State<HomeScreen> {
               ),
               const SizedBox(height: 50),
               Row(
+                mainAxisSize: MainAxisSize.min,
                 mainAxisAlignment: MainAxisAlignment.spaceAround,
                 children: [
                   Expanded(
