+++ lib/screens/launch_screen.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
-import 'package:intern_task/pages/intro_screen.dart';
+import 'package:intern_task/screens/intro_screen.dart';
 
 class LaunchScreen extends StatefulWidget {
   const LaunchScreen({super.key});
