+++ lib/screens/myaccount_screen.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:intern_task/pages/menu_button.dart';
+import 'package:intern_task/util/widget/menu_button.dart';
 
 class MyAccountScreen extends StatefulWidget {
   const MyAccountScreen({super.key});
@@ -21,6 +21,7 @@ class _MyAccountScreenState extends State<MyAccountScreen> {
                 },
                 icon: const Icon(Icons.arrow_back_ios_sharp))),
         body: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
           children: [
             const SizedBox(height: 30),
             Expanded(
@@ -69,14 +70,7 @@ class _MyAccountScreenState extends State<MyAccountScreen> {
                 child: Column(
                   crossAxisAlignment: CrossAxisAlignment.start,
                   children: [
-                    const MenuButton(buttonName: 'Account Setting'),
-                    const MenuButton(buttonName: 'Download Options'),
-                    const MenuButton(buttonName: 'Notifications Setting'),
-                    const SizedBox(height: 20),
-                    const MenuButton(buttonName: 'Privacy & Policy'),
-                    const MenuButton(buttonName: 'Help Center'),
-                    const MenuButton(buttonName: 'About Us'),
-                    const SizedBox(height: 10),
+                    ...menulist,
                     TextButton(
                       onPressed: () {},
                       child: const Text(
@@ -88,17 +82,16 @@ class _MyAccountScreenState extends State<MyAccountScreen> {
                         ),
                       ),
                     ),
-                    const Expanded(child: SizedBox()),
-                    ElevatedButton(
-                        onPressed: () {
-                          Navigator.pushNamed(context, '/login');
-                        },
-                        child: const Text('Log Out')),
-                    const Expanded(child: SizedBox()),
                   ],
                 ),
               ),
             ),
+            ElevatedButton(
+                onPressed: () {
+                  Navigator.pushNamed(context, '/login');
+                },
+                child: const Text('Log Out')),
+            const Expanded(child: SizedBox()),
           ],
         ),
       ),
