+++ pubspec.yaml
@@ -36,6 +36,8 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
+  firebase_auth: ^4.1.1
+  firebase_core: ^2.1.1
 
 dev_dependencies:
   flutter_test:
